# src/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This API provides all the functions needed to manage \nthe JABC HR\
    \ system.\n"
  version: "1.0.0"
  title: "JABC"
  contact:
    email: "jabraham9719@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/JABC/1.0.0"
...
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /performance/{id}:
    get:
      tags:
      - "Performance"
      summary: "gets an specific Performance"
      description: "Will return the Performance that matches with the provided [id]\
        \ from \nthe Employee with [id]\n"
      operationId: "getPerformance"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the searched Performance"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IPerformance"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Performance"
    put:
      tags:
      - "Performance"
      summary: "updates the Performance"
      description: "Will update an Performance with the provided data in body \nif\
        \ the Performance matches the [id]\n"
      operationId: "updatePerformance"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the searched Performance"
        required: true
        type: "integer"
        format: "uint64"
      - in: "body"
        name: "performance"
        description: "Performance data"
        required: true
        schema:
          $ref: "#/definitions/IPerformance"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Performance"
    delete:
      tags:
      - "Performance"
      summary: "deletes Performance"
      description: "Will delete the Performance that matches with the provided [id]\
        \ from \nthe Employee with [id]\n"
      operationId: "deletePerformance"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the searched Performance"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Performance"
  /performance/{id}/comment:
    get:
      tags:
      - "Performance"
      summary: "get all the Comments from a Performance with [id]"
      description: "This returns all the Comments of the Performance with [id].\n"
      operationId: "getComments"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the Performance with comments"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/IComment"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Performance"
    post:
      tags:
      - "Performance"
      summary: "creates a new Comment for the Performance with [id]"
      description: "Will create a new Comment with the provided data in body, to the\
        \ Performance with [id]"
      operationId: "createComment"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the Performance to be commented"
        required: true
        type: "integer"
        format: "uint64"
      - in: "body"
        name: "comment"
        description: "Comment data"
        required: true
        schema:
          $ref: "#/definitions/IComment"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Performance"
  /performance/{id}/comment/{idComment}:
    get:
      tags:
      - "Performance"
      summary: "gets an specific Comment"
      description: "Will return the Comment that matches with the provided [idComment]"
      operationId: "getComment"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the Performance with comments"
        required: true
        type: "integer"
        format: "uint64"
      - name: "idComment"
        in: "path"
        description: "idComment of the searched Comment"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IComment"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Performance"
    put:
      tags:
      - "Performance"
      summary: "updates the Comment"
      description: "Will update an Comment with the provided data in body if the Comment\
        \ matches the [idComment]"
      operationId: "updateComment"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the Performance with comments"
        required: true
        type: "integer"
        format: "uint64"
      - name: "idComment"
        in: "path"
        description: "idComment of the Comment to be updated"
        required: true
        type: "integer"
        format: "uint64"
      - in: "body"
        name: "comment"
        description: "Comment data"
        required: true
        schema:
          $ref: "#/definitions/IComment"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Performance"
    delete:
      tags:
      - "Performance"
      summary: "deletes Comment"
      description: "Will delete an Comment if the Comment matches the [idComment]"
      operationId: "deleteComment"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "Auth Token that grants access to the system"
        required: false
        type: "string"
      - name: "id"
        in: "path"
        description: "id of the Performance with comments"
        required: true
        type: "integer"
        format: "uint64"
      - name: "idComment"
        in: "path"
        description: "idComment of the Comment to be deleted"
        required: true
        type: "integer"
        format: "uint64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/IApiResponse"
        401:
          description: "unauthorized access"
          schema:
            $ref: "#/definitions/IApiResponse"
        default:
          description: "bad request"
          schema:
            $ref: "#/definitions/IApiResponse"
      security:
      - AuthToken: []
      x-swagger-router-controller: "Performance"
...
securityDefinitions:
  AuthToken:
    type: "apiKey"
    name: "X-Auth-Token"
    in: "header"
definitions:
  IApiResponse:
    type: "object"
    required:
    - "message"
    - "responseCode"
    - "type"
    properties:
      responseCode:
        type: "integer"
        format: "int32"
      type:
        type: "string"
        enum:
        - "ERROR"
        - "SUCCESS"
      message:
        type: "string"
        example: "Unauthorized access to the API"
      debugMessage:
        type: "string"
        example: "This is a debug message"
    example:
      debugMessage: "This is a debug message"
      type: "ERROR"
      message: "Unauthorized access to the API"
      responseCode: 0
...
  IPerformance:
    type: "object"
    required:
    - "date"
    - "fkEmployee"
    - "id"
    - "status"
    properties:
      id:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The unique identifier of the Performance"
      fkEmployee:
        type: "integer"
        format: "uint64"
        example: 1
        description: "Foreign key of the employee with this performance review"
      date:
        type: "number"
        format: "uint64"
        description: "The unix timestamp of the date the performance review was created"
      status:
        type: "integer"
      jabcGoals:
        type: "array"
        description: "Contains all the JABC Goals (Section 1) of this performance\
          \ review"
        items:
          $ref: "#/definitions/IJABCGoal"
      personalTargets:
        type: "array"
        description: "Contains all the Personal Targets (Section 2) of this performance\
          \ review"
        items:
          $ref: "#/definitions/IPersonalTarget"
      objectives:
        type: "array"
        description: "Contains all the Objectives (Section 3) of this performance\
          \ review"
        items:
          $ref: "#/definitions/IObjective"
      developmentGoals:
        type: "array"
        description: "Contains all the Development Goals (Section 4) of this performance\
          \ review"
        items:
          $ref: "#/definitions/IDevelopmentGoal"
      comments:
        type: "array"
        description: "Contains all the comments (Section 5) of this performance review"
        items:
          $ref: "#/definitions/IComment"
    description: "A Performance contains info of a performance record of an employee\n"
    example:
      date: 0.80082819046101150206595775671303272247314453125
      personalTargets:
      - rating: "rating"
        description: "description"
        fkPerformance: 1
        id: 1
      - rating: "rating"
        description: "description"
        fkPerformance: 1
        id: 1
      comments:
      - fkCommenter: 1
        date: 1
        fkPerformance: 1
        comment: "comment"
        id: 1
      - fkCommenter: 1
        date: 1
        fkPerformance: 1
        comment: "comment"
        id: 1
      fkEmployee: 1
      objectives:
      - q1: "q1"
        innovative: "innovative"
        q2: "q2"
        q3: "q3"
        q4: "q4"
        impact: "impact"
        fkPerformance: 1
        id: 1
        foundation: "foundation"
        volAlum: "volAlum"
        relevance: "relevance"
      - q1: "q1"
        innovative: "innovative"
        q2: "q2"
        q3: "q3"
        q4: "q4"
        impact: "impact"
        fkPerformance: 1
        id: 1
        foundation: "foundation"
        volAlum: "volAlum"
        relevance: "relevance"
      id: 1
      jabcGoals:
      - goal: "goal"
        name: "name"
        fkPerformance: 1
        id: 1
        previousYear: "previousYear"
      - goal: "goal"
        name: "name"
        fkPerformance: 1
        id: 1
        previousYear: "previousYear"
      status: 6
      developmentGoals:
      - goal: "goal"
        keyActivities: "keyActivities"
        rating: "rating"
        fkPerformance: 1
        id: 1
      - goal: "goal"
        keyActivities: "keyActivities"
        rating: "rating"
        fkPerformance: 1
        id: 1
  IJABCGoal:
    type: "object"
    required:
    - "fkPerformance"
    - "id"
    properties:
      id:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The unique identifier of the JABCGoal"
      fkPerformance:
        type: "integer"
        format: "uint64"
        example: 1
        description: "Foreign key of Performance that has this JABCGoal"
      name:
        type: "string"
      goal:
        type: "string"
      previousYear:
        type: "string"
    description: "A JABCGoal contains what another employee thinks about the performance\
      \ record of another employee\n"
    example:
      goal: "goal"
      name: "name"
      fkPerformance: 1
      id: 1
      previousYear: "previousYear"
  IPersonalTarget:
    type: "object"
    required:
    - "fkPerformance"
    - "id"
    properties:
      id:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The unique identifier of the PersonalTarget"
      fkPerformance:
        type: "integer"
        format: "uint64"
        example: 1
        description: "Foreign key of Performance that has this PersonalTarget"
      description:
        type: "string"
      rating:
        type: "string"
    description: "A PersonalTarget contains what another employee thinks about the\
      \ performance record of another employee\n"
    example:
      rating: "rating"
      description: "description"
      fkPerformance: 1
      id: 1
  IObjective:
    type: "object"
    required:
    - "fkPerformance"
    - "id"
    properties:
      id:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The unique identifier of the Objective"
      fkPerformance:
        type: "integer"
        format: "uint64"
        example: 1
        description: "Foreign key of Performance that has this Objective"
      q1:
        type: "string"
      q2:
        type: "string"
      q3:
        type: "string"
      q4:
        type: "string"
      impact:
        type: "string"
      relevance:
        type: "string"
      volAlum:
        type: "string"
      innovative:
        type: "string"
      foundation:
        type: "string"
    description: "A Objective contains what another employee thinks about the performance\
      \ record of another employee\n"
    example:
      q1: "q1"
      innovative: "innovative"
      q2: "q2"
      q3: "q3"
      q4: "q4"
      impact: "impact"
      fkPerformance: 1
      id: 1
      foundation: "foundation"
      volAlum: "volAlum"
      relevance: "relevance"
  IDevelopmentGoal:
    type: "object"
    required:
    - "fkPerformance"
    - "id"
    properties:
      id:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The unique identifier of the DevelopmentGoal"
      fkPerformance:
        type: "integer"
        format: "uint64"
        example: 1
        description: "Foreign key of Performance that has this DevelopmentGoal"
      goal:
        type: "string"
      keyActivities:
        type: "string"
      rating:
        type: "string"
    description: "A DevelopmentGoal contains what another employee thinks about the\
      \ performance record of another employee\n"
    example:
      goal: "goal"
      keyActivities: "keyActivities"
      rating: "rating"
      fkPerformance: 1
      id: 1
  IComment:
    type: "object"
    required:
    - "comment"
    - "date"
    - "fkCommenter"
    - "fkPerformance"
    - "id"
    properties:
      id:
        type: "integer"
        format: "uint64"
        example: 1
        description: "The unique identifier of the Comment"
      fkPerformance:
        type: "integer"
        format: "uint64"
        example: 1
        description: "Foreign key of Performance that has this Comment"
      fkCommenter:
        type: "integer"
        format: "uint64"
        example: 1
        description: "Foreign key of Employee that created the Comment"
      comment:
        type: "string"
      date:
        type: "integer"
        format: "uint64"
        description: "The unix timestamp of the date when the Comment was created"
    description: "A Comment contains what another employee thinks about the performance\
      \ record of another employee\n"
    example:
      fkCommenter: 1
      date: 1
      fkPerformance: 1
      comment: "comment"
      id: 1
...